requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
line-bot-sdk==2.4.2
google-generativeai==0.3.2
python-dotenv==1.0.1
loguru==0.7.2